+++ shared-libs/crates/start-core/src/install/mod.rs
@@ -519,7 +519,7 @@ pub async fn cli_install(
                 &method.join("."),
                 to_value(&InstallParams {
                     id,
-                    registry: ctx.registry_url.clone().or_not_found("--registry")?,
+                    registry: ctx.registry_url()?.or_not_found("--registry")?,
                     version,
                 })?,
             )