+++ shared-libs/crates/start-core/src/registry/context.rs
@@ -205,7 +205,7 @@ impl CallRemote<RegistryContext> for CliContext {
     ) -> Result<Value, RpcError> {
         let local_auth = local_auth_header::<RegistryContext>().await;
 
-        let url = if let Some(url) = self.registry_url.clone() {
+        let url = if let Some(url) = self.registry_url()? {
             url
         } else if local_auth.is_some() || !self.registry_hostname.is_empty() {
             let mut url: Url = format!(